backend/requirements.txt
# 基础框架
pydantic>=2.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
loguru>=0.7.0

# 数据库
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 邮箱验证
email-validator>=2.0

# AI 模型调用
openai>=1.0.0
httpx>=0.24.0

# 向量计算
numpy>=1.24.0

# 文件上传
python-multipart>=0.0.6
